fix(gameserver): round float intermediates explicitly so Unity and .NET agree
The golden vectors failed 3 of 96 cases on their first run inside Unity. All
three traced to one shape: x * x + y * y, in Vec2.SqrMagnitude and in
MovementSystem.ResolveDirection.

C# permits a float expression to be evaluated at higher precision than float
(ECMA-334 11.3.7). .NET 10's RyuJIT evaluates strictly in float32; Unity's Editor
Mono JIT keeps double-precision intermediates and rounds once at the end. Both
conform. The results differ by one ULP — and because that value feeds the
deadzone and magnitude-clamp comparisons, the two runtimes could take different
branches, not merely report slightly different numbers.

Every arithmetic intermediate in Vec2 and MovementSystem is now cast to float per
operation. Integrate additionally splits its multiply-add into a separate float
local: a + b * c can be contracted into a single FMA that rounds once instead of
twice, giving a third possible answer.

The server's own results are unchanged — RyuJIT already evaluated in float32, so
the casts are a no-op there and every existing golden vector still passes. The
fix moves Unity onto the server's answer rather than the reverse.

ADR-10 rule 5 is amended: choosing IEEE-exact operations was necessary but not
sufficient, and the rule now covers intermediate rounding and FMA contraction.
Worth recording how this surfaced — the operations were already legal, the whole
server suite passed, and nothing warned. Only replaying the vectors under the
other runtime exposed it, which is the entire reason that gate exists.

+++ backend/gameserver-dotnet/Shared.GameLogic/Components/Vec2.cs
@@ -28,8 +28,20 @@ namespace Shared.GameLogic.Components
 
         // --- Properties ---
 
-        /// <summary>Squared magnitude (avoids sqrt).</summary>
-        public float SqrMagnitude => X * X + Y * Y;
+        /// <summary>
+        /// Squared magnitude (avoids sqrt).
+        /// </summary>
+        /// <remarks>
+        /// Each intermediate is cast to <c>float</c> explicitly. C# permits a
+        /// float expression to be evaluated at higher precision (ECMA-334
+        /// §11.3.7), and the two runtimes this code has to agree on take
+        /// different options: .NET 10's RyuJIT evaluates strictly in float32,
+        /// while Unity's Editor Mono JIT keeps double-precision intermediates
+        /// and rounds once at the end. Written as <c>X * X + Y * Y</c> the two
+        /// disagreed by one ULP, which the golden vectors caught.
+        /// Do not "simplify" this back to a single expression.
+        /// </remarks>
+        public float SqrMagnitude => (float)((float)(X * X) + (float)(Y * Y));
 
         /// <summary>Magnitude (length) of the vector.</summary>
         public float Magnitude => MathF.Sqrt(SqrMagnitude);
@@ -49,9 +61,10 @@ namespace Shared.GameLogic.Components
         /// <summary>Squared Euclidean distance (no sqrt — use for comparisons).</summary>
         public static float DistanceSq(in Vec2 a, in Vec2 b)
         {
-            float dx = a.X - b.X;
-            float dy = a.Y - b.Y;
-            return dx * dx + dy * dy;
+            // Per-operation float casts, for the reason given on SqrMagnitude.
+            float dx = (float)(a.X - b.X);
+            float dy = (float)(a.Y - b.Y);
+            return (float)((float)(dx * dx) + (float)(dy * dy));
         }
 
         /// <summary>Euclidean distance between two points.</summary>
